notify: add a channel to the home assistant notification
https://companion.home-assistant.io/docs/notifications/notifications-basic/#notification-channels

This passes through to be a specific notification category on Android (I'm not sure how iOS handles these).

In the future the notification type (i.e. motion, person, etc) could be appended to the channel name to allow even more specific notification tuning. I didn't see this data in the `NotifierOptions` object passed in.